consolidate sentry apps web
