+++ src/sentry/sentry_apps/web/sentryapp_avatar.py
