+++ tests/sentry/sentry_apps/web/test_sentryapp_avatar.py
